#33 wip split stores directory
