+++ src/hooks/useQueryParams.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useMemo } from 'react';
-import { usePersistQueryParams } from '../stores/currency/persistQueryParamsStore';
+import { usePersistQueryParams } from '../features/currency/stores/persistQueryParamsStore';
 import type { QueryParams } from '../types';
 import { parseSearch, buildSearch } from '../utils';
 import { useLocationSearch } from './useLocationSearch';
