+++ src/pages/ConvertPage/ConvertPage.tsx
@@ -5,12 +5,12 @@ import { Converter } from '../../features/currency/components/Converter';
 import { EmptyFragment } from '../../features/currency/components/EmptyFragment';
 import { ErrorsFragment } from '../../features/currency/components/ErrorsFragment';
 import { SymbolChips } from '../../features/currency/components/SymbolChips';
-import { extractCurrencyCodes } from '../../features/currency/utils';
+import { useCurrencies } from '../../features/currency/stores/currenciesStore';
 import type {
   CodeString,
   Currency,
 } from '../../features/currency/types';
-import { useCurrencies } from '../../stores/currency/currenciesStore';
+import { extractCurrencyCodes } from '../../features/currency/utils';
 
 export type ConvertersData = {
   rates: {
